+++ odoo/local-src/specific_helpdesk/__manifest__.py
@@ -1,0 +1,17 @@
+# -*- coding: utf-8 -*-
+# Copyright 2017 Camptocamp SA
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+{'name': 'Specifc Helpdesk for Alcyon',
+ 'version': '10.0.1.0.0',
+ 'author': 'Camptocamp',
+ 'license': 'AGPL-3',
+ 'category': 'Helpdesk',
+ 'depends': ['helpdesk',
+             ],
+ 'website': 'http://www.camptocamp.com',
+ 'data': [
+     'security/ir.model.access.csv',
+     'views/helpdesk.xml',
+ ],
+ 'installable': True,
+ }