+++ odoo/local-src/specific_helpdesk/models/helpdesk.py
@@ -1,0 +1,26 @@
+# -*- coding: utf-8 -*-
+# Copyright 2017 Camptocamp SA
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from odoo import models, fields
+from odoo.addons.base.res.res_request import referenceable_models
+
+
+class HelpdeskTicket(models.Model):
+
+    _inherit = 'helpdesk.ticket'
+
+    helpdesk_ticket_reason_id = fields.Many2one(
+        comodel_name='helpdesk.ticket.reason', string='Reason')
+
+    ref = fields.Reference(
+        selection=referenceable_models,
+        string='Reference')
+
+
+class HelpdeskTicketReason(models.Model):
+
+    _name = 'helpdesk.ticket.reason'
+    _description = 'Ticket Reason'
+
+    name = fields.Char(string='Name')