+++ odoo/local-src/specific_product/__manifest__.py
@@ -17,6 +17,7 @@
     'data': [
         'views/product_pricelist.xml',
         'views/product_template.xml',
+        'views/product_product.xml',
     ],
     'installable': True,
 }