+++ HISTORY.rst
@@ -15,6 +15,8 @@ Unreleased
 * ALCN-820: Add specific module to shipping costs calculation
 * ALCN-836: Add 'Sale Order Lines Unavailable' menu in Sales > Report
 
+* ALCN-847: Add CNK field to product template
+
 **Bugfixes**
 
 **Build**